Move `lookupDeployIDFromTaskDefinitionARN()` function and tests into `common` package
