+++ api/provider/aws/common_test.go
@@ -1,0 +1,36 @@
+package aws
+
+import (
+	"testing"
+
+	"github.com/quintilesims/layer0/api/tag"
+	"github.com/quintilesims/layer0/common/models"
+	"github.com/stretchr/testify/assert"
+)
+
+func Test_lookupDeployIDFromTaskDefinitionARN(t *testing.T) {
+	tagStore := tag.NewMemoryStore()
+	service := NewServiceProvider(nil, tagStore, nil)
+
+	tags := models.Tags{
+		{
+			EntityID:   "dpl_id",
+			EntityType: "deploy",
+			Key:        "arn",
+			Value:      "task_definition_arn",
+		},
+	}
+
+	for _, tag := range tags {
+		if err := tagStore.Insert(tag); err != nil {
+			t.Fatal(err)
+		}
+	}
+
+	result, err := service.lookupDeployIDFromTaskDefinitionARN("task_definition_arn")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	assert.Equal(t, "dpl_id", result)
+}
